refactor(api): move event indexer into internal/stellar; feat: yield forecasting
#454 — extract the ~420-line event-indexer (background loop, Soroban fetch,
event apply, cursor, dedup) out of cmd/api/main.go into a cohesive, testable
internal/stellar package: exported StartEventIndexer and EventSyncer
(SyncEvents), with the logic and its tests now living beside the other stellar
code instead of in main. main.go just wires it. (Full repository-layer
migration left as follow-up — repo methods are UUID-keyed while events are
contract-address-keyed.)

#118 — add internal/forecast: an explainable yield-forecasting model that
blends a recency-weighted APY average with TVL-trend, stablecoin-rate, and
policy-rate signals, returning a predicted APY, a volatility-based confidence
band, a confidence score, and a rationale. Fully unit-tested.
